Refactored OOD into its own file
